backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
google-auth
requests
google-generativeai